Add linux, dhcpcd, and netctl to the base installation
The base group was changed to a base meta package and many packages were
removed: https://www.archlinux.org/news/base-group-replaced-by-mandatory-base-package-manual-intervention-required/

Therefore, it's necessary to add linux or linux-lts to install a Kernel.
Dhcpcd is needed to get an IP after the first reboot. Netctl is needed later when
using Vagrant.